+++ example/lib/special_text/image_text.dart
@@ -4,7 +4,6 @@ import 'package:extended_image/extended_image.dart';
 import 'package:extended_text_field/extended_text_field.dart';
 import 'package:flutter/material.dart';
 import 'package:html/parser.dart';
-import 'my_special_text_span_builder.dart';
 
 class ImageText extends SpecialText {
   static const String flag = "<img";